CSS: deprecate jQuery.cssProps
Fixes gh-3653
