+++ src/css/var/cssProps.js
@@ -1,0 +1,8 @@
+define( function() {
+	"use strict";
+
+	// Add in properties whose names you wish to fix before
+	// setting or getting the value
+	// Also used for the vendor prefix cache
+	return {};
+} );
